Changed Flask routes from post / render / routes classes to separate blueprints
